backend/src/requirements.txt
asgiref==3.3.4
Babel==2.9.0
backports.entry-points-selectable==1.1.0
certifi==2019.11.28
cffi==1.14.4
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.3.1
defusedxml==0.6.0
distlib==0.3.2
dj-rest-auth==2.1.11
Django==3.2.6
django-allauth==0.43.0
django-cors-headers==3.6.0
django-resized==0.3.11
django-rest-knox==4.1.0
djangorestframework==3.12.4
djangorestframework-simplejwt==4.7.1
drf-yasg==1.20.0
filelock==3.0.12
idna==2.9
inflection==0.5.1
install==1.3.4
itypes==1.2.0
Jinja2==3.0.1
MarkupSafe==2.0.1
mysqlclient==2.0.3
oauthlib==3.1.0
packaging==21.0
Pillow==8.3.1
platformdirs==2.2.0
pycparser==2.20
PyJWT==2.1.0
pyparsing==2.4.7
python-dateutil==2.8.1
python-mimeparse==1.6.0
python3-openid==3.1.0
pytz==2020.5
requests==2.23.0
requests-oauthlib==1.3.0
ruamel.yaml==0.17.10
ruamel.yaml.clib==0.2.6
six==1.15.0
sqlparse==0.4.1
typing-extensions==3.10.0.0
uritemplate==3.0.1
urllib3==1.25.8
virtualenv==20.7.2
